requirements.txt
django==1.9.2 
djangorestframework==3.3.2
django-filter
django-rest-swagger
psycopg2
requests

sniffer
django-jenkins
django-extensions
coverage
responses
mock
markdown

## testing:
nose
django_nose
pinocchio
nose-html-reporting

gunicorn

